% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{autoslider.trade-package}
\alias{autoslider.trade}
\alias{autoslider.trade-package}
\title{autoslider.trade Package}
\description{
Trading tables, listings and figures, downstream of \code{autoslider.core}.
Outputs follow the \code{autoslider} naming convention: \verb{t_*_slide()} for tables,
\verb{l_*_slide()} for listings and \verb{g_*_slide()} for figures.
}
\author{
\strong{Maintainer}: Joe Zhu \email{sha.joe.zhu@gmail.com}

Authors:
\itemize{
  \item Joe Zhu \email{sha.joe.zhu@gmail.com}
}

}
